wip: poc: client higher level api
